Move counter.dart to example.dart for pub.dev visibility
